requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.3
python-dotenv==1.0.1
openai==1.51.2
azure-search-documents==11.5.1
azure-storage-blob==12.23.1